Refactor
Update variable names, fix formatting issues, and move QExample and
QOption into their own files.
