+++ src/qexample.cpp
@@ -1,0 +1,14 @@
+#include "qexample.h"
+
+#include <QJsonObject>
+
+QJsonObject QExample::toJson() const
+{
+    QJsonObject object
+    {
+        {"example", example},
+        {"description", description},
+    };
+
+    return object;
+}
